Oxygen Calibration Analysis
I have complied the work me, Blandine and Paul did on 5th of May and written comments and descriptions on how we used the Winkler titration. Basically the final result is that the values are too uncertain and there is no obvious correlation so we will not perform a calibration.